Moved Chat dataclass to separate models.py
